Rename MetricsDataSourceWriter to MetricsOnlineDatabaseWriter
