+++ src/main/kotlin/com/zepben/evolve/database/online/MetricsOnlineDatabaseWriter.kt
@@ -1,38 +1,42 @@
 /*
- * Copyright 2024 Zeppelin Bend Pty Ltd
+ * Copyright 2025 Zeppelin Bend Pty Ltd
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package com.zepben.evolve.database.sqlite.metrics
+package com.zepben.evolve.database.online
 
 import com.zepben.evolve.database.sqlite.common.SchemaUtils
 import com.zepben.evolve.database.sqlite.common.TableVersion
 import com.zepben.evolve.database.sqlite.extensions.configureBatch
+import com.zepben.evolve.database.sqlite.metrics.MetricsDatabaseTables
+import com.zepben.evolve.database.sqlite.metrics.MetricsWriter
 import com.zepben.evolve.metrics.IngestionJob
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.sql.Connection
-import javax.sql.DataSource
 
 /**
- * Class for writing metrics to an arbitrary datasource.
+ * Class for writing metrics to an online SQL database.
  */
-class MetricsDataSourceWriter internal constructor(
-    private val dataSource: DataSource,
+class MetricsOnlineDatabaseWriter internal constructor(
+    private val getConnection: () -> Connection,
     private val databaseTables: MetricsDatabaseTables,
     private val schemaUtils: SchemaUtils = SchemaUtils(databaseTables),
     private val createMetricsWriter: (IngestionJob, MetricsDatabaseTables) -> MetricsWriter = { job, tables -> MetricsWriter(job, tables) }
 ) {
 
-    constructor(dataSource: DataSource) : this(dataSource, MetricsDatabaseTables())
+    /**
+     * @param getConnection Provider of the connection to the specified database.
+     */
+    constructor(getConnection: () -> Connection) : this(getConnection, MetricsDatabaseTables())
 
     private val logger: Logger = LoggerFactory.getLogger(javaClass)
     private val versionTable = databaseTables.getTable<TableVersion>()
 
-    fun save(job: IngestionJob): Boolean = dataSource.connection.use { connection ->
+    fun save(job: IngestionJob): Boolean = getConnection().use { connection ->
         connection.configureBatch()
         val localVersion = versionTable.supportedVersion
         val status = when (val remoteVersion = schemaUtils.getVersion(connection)) {
